Move decorators in app/decorators path
